// src/app/core/models/driver/driver-add.model.ts
import{DeviceManagmentAddModel} from 'src/app/core/models/driver/device-managment-add.model';
export class DiverAddModel {
    firstName: string;
    lastName: string;
    email: string;
    address: string;
    joinedDate: string;
    phoneNumber: string;
    licenceNumber: string;
    vehicleType: string;
    insurenceProvider: string;
    policyNumber :string;
    active: string;
    createDateTime: string;
    updateDateTime:string;
    gstnumber: string
    isActive:string;
    device:DeviceManagmentAddModel;
}